refactor: split dashboard sections into dynamic components
